add ens convo templates, localization
