+++ packages/easywallet/src/conversations/ens/ens.ts
@@ -1,0 +1,10 @@
+import { MyContext, MyConversation } from "../../context.js";
+
+/**
+ * ENS Menu
+ * @param conversation
+ * @param ctx
+ */
+export async function ens(_conversation: MyConversation, ctx: MyContext) {
+    await ctx.reply("ens", { parse_mode: "Markdown" });
+}
