+++ packages/easywallet/src/conversations/index.ts
@@ -1,3 +1,4 @@
+export * from "./ens/index.js";
 export * from "./intro.js";
 export * from "./start.js";
 export * from "./wallet.js";
