+++ packages/easywallet/src/templates/index.ts
@@ -1,1 +1,2 @@
 export * from "./getIntroMessage.js";
+export * from "./loadTemplate.js";
