+++ packages/easywallet/templates/en/intro.md
@@ -1,4 +1,4 @@
-👋 *Welcome to BOT_NAME!* 💸
+👋 *Welcome to $BOT_NAME!* 💸
 A simple crypto wallet in Telegram. It's the wallet grandma can use 😉
 
 ⭐️ *Features*
@@ -6,6 +6,6 @@ A simple crypto wallet in Telegram. It's the wallet grandma can use 😉
 ⛽️ *Gasless*: No blockchain fees powered by ERC4337 Account Abstraction
 
 💫 *Tips*
-⚠️ BOT_NAME is still experimental and in beta. Avoid depositing too much crypto and *use at your own risk*!
+⚠️ $BOT_NAME is still experimental and in beta. Avoid depositing too much crypto and *use at your own risk*!
 
 To view this message again, type /intro
