+++ packages/easywallet/templates/fr/intro.md
@@ -1,0 +1,11 @@
+👋 *Bienvenue sur $BOT_NAME !* 💸
+Un simple portefeuille crypto sur Telegram. C'est le portefeuille que même grand-mère peut utiliser 😉
+
+⭐️ *Caractéristiques*
+🔑 *Dépositaire*: La technologie de calcul multipartite (MPC) évite la nécessité de stocker une clé privée. (auto-détention à venir bientôt !)
+⛽️ *Sans frais de gaz*: Pas de frais de blockchain grâce à l'abstraction de compte ERC4337
+
+💫 *Conseils*
+⚠️ $BOT_NAME est encore expérimental et en version bêta. Évitez de déposer trop de crypto et *utilisez à vos propres risques* !
+
+Pour revoir ce message, tapez /intro
